requirements.txt
requests
PyGithub
beautifulsoup4
playwright
